Extract ICMP test helpers into nym-ip-packet-requests   - Add icmp_utils module behind test-utils feature flag   - Move build_icmp_ping, build_icmpv6_ping, is_echo_reply_v4/v6 from     example   - Update ipr_tunnel example to use shared helpers
