+++ sdk/rust/nym-sdk/examples/ipr_tunnel.rs
@@ -14,14 +14,10 @@ use std::net::{Ipv4Addr, Ipv6Addr};
 use std::time::Duration;
 
 use nym_ip_packet_requests::codec::MultiIpPacketCodec;
+use nym_ip_packet_requests::icmp_utils::{
+    build_icmp_ping, build_icmpv6_ping, is_echo_reply_v4, is_echo_reply_v6,
+};
 use nym_sdk::ipr_wrapper::IpMixStream;
-use pnet_packet::icmp::echo_reply::EchoReplyPacket;
-use pnet_packet::icmp::echo_request::MutableEchoRequestPacket;
-use pnet_packet::icmp::{IcmpPacket, IcmpTypes};
-use pnet_packet::icmpv6::Icmpv6Types;
-use pnet_packet::ipv4::{Ipv4Flags, MutableIpv4Packet};
-use pnet_packet::ipv6::MutableIpv6Packet;
-use pnet_packet::Packet;
 
 const PING4_TARGET: Ipv4Addr = Ipv4Addr::new(8, 8, 8, 8);
 const PING6_TARGET: Ipv6Addr = Ipv6Addr::new(0x2001, 0x4860, 0x4860, 0, 0, 0, 0, 0x8888);
@@ -49,13 +45,13 @@ async fn main() -> anyhow::Result<()> {
     println!("Tunnel up — IPv4: {src4}, IPv6: {src6}");
 
     // Send IPv4 ping (ICMP seq=0, unrelated to LP Stream sequence numbers)
-    let pkt4 = build_icmp_ping(src4, PING4_TARGET, 0)?;
+    let pkt4 = build_icmp_ping(src4, PING4_TARGET, 0).expect("failed to build ICMP packet");
     let bundled = MultiIpPacketCodec::bundle_one_packet(pkt4.into());
     tunnel.send_ip_packet(&bundled).await?;
     println!("Sent ping → {PING4_TARGET}");
 
     // Send IPv6 ping
-    let pkt6 = build_icmpv6_ping(src6, PING6_TARGET, 0)?;
+    let pkt6 = build_icmpv6_ping(src6, PING6_TARGET, 0).expect("failed to build ICMPv6 packet");
     let bundled = MultiIpPacketCodec::bundle_one_packet(pkt6.into());
     tunnel.send_ip_packet(&bundled).await?;
     println!("Sent ping → {PING6_TARGET}");
@@ -93,129 +89,4 @@ async fn main() -> anyhow::Result<()> {
 
     tunnel.disconnect().await;
     Ok(())
-}
-
-// --- IPv4 ICMP helpers ---
-
-fn build_icmp_ping(src: Ipv4Addr, dst: Ipv4Addr, seq: u16) -> anyhow::Result<Vec<u8>> {
-    let mut echo = MutableEchoRequestPacket::owned(vec![0u8; 64])
-        .ok_or_else(|| anyhow::anyhow!("failed to create ICMP packet"))?;
-    echo.set_icmp_type(IcmpTypes::EchoRequest);
-    echo.set_icmp_code(pnet_packet::icmp::IcmpCode::new(0));
-    echo.set_sequence_number(seq);
-    let cksum = pnet_packet::icmp::checksum(
-        &IcmpPacket::new(echo.packet()).ok_or_else(|| anyhow::anyhow!("checksum failed"))?,
-    );
-    echo.set_checksum(cksum);
-
-    let total_len = 20 + echo.packet().len();
-    let mut ip = MutableIpv4Packet::owned(vec![0u8; total_len])
-        .ok_or_else(|| anyhow::anyhow!("failed to create IPv4 packet"))?;
-    ip.set_version(4);
-    ip.set_header_length(5);
-    ip.set_total_length(total_len as u16);
-    ip.set_ttl(64);
-    ip.set_next_level_protocol(pnet_packet::ip::IpNextHeaderProtocols::Icmp);
-    ip.set_source(src);
-    ip.set_destination(dst);
-    ip.set_flags(Ipv4Flags::DontFragment);
-    ip.set_payload(echo.packet());
-
-    let mut buf = ip.consume_to_immutable().packet().to_vec();
-    let cksum = ipv4_checksum(&buf);
-    buf[10] = (cksum >> 8) as u8;
-    buf[11] = cksum as u8;
-    Ok(buf)
-}
-
-fn ipv4_checksum(header: &[u8]) -> u16 {
-    let mut sum = 0u32;
-    for i in (0..20).step_by(2) {
-        sum += ((header[i] as u32) << 8) | header[i + 1] as u32;
-    }
-    while (sum >> 16) > 0 {
-        sum = (sum & 0xFFFF) + (sum >> 16);
-    }
-    !sum as u16
-}
-
-fn is_echo_reply_v4(data: &[u8], expected_dst: Ipv4Addr) -> bool {
-    let Some(ip) = pnet_packet::ipv4::Ipv4Packet::new(data) else {
-        return false;
-    };
-    if ip.get_destination() != expected_dst {
-        return false;
-    }
-    if ip.get_next_level_protocol() != pnet_packet::ip::IpNextHeaderProtocols::Icmp {
-        return false;
-    }
-    let Some(reply) = EchoReplyPacket::new(ip.payload()) else {
-        return false;
-    };
-    if reply.get_icmp_type() != IcmpTypes::EchoReply {
-        return false;
-    }
-    println!(
-        "  IPv4 reply: {} → {}, seq={}",
-        ip.get_source(),
-        ip.get_destination(),
-        reply.get_sequence_number(),
-    );
-    true
-}
-
-// --- IPv6 ICMPv6 helpers ---
-
-fn build_icmpv6_ping(src: Ipv6Addr, dst: Ipv6Addr, seq: u16) -> anyhow::Result<Vec<u8>> {
-    let mut echo =
-        pnet_packet::icmpv6::echo_request::MutableEchoRequestPacket::owned(vec![0u8; 64])
-            .ok_or_else(|| anyhow::anyhow!("failed to create ICMPv6 packet"))?;
-    echo.set_icmpv6_type(Icmpv6Types::EchoRequest);
-    echo.set_icmpv6_code(pnet_packet::icmpv6::Icmpv6Code::new(0));
-    echo.set_sequence_number(seq);
-    let cksum = pnet_packet::icmpv6::checksum(
-        &pnet_packet::icmpv6::Icmpv6Packet::new(echo.packet())
-            .ok_or_else(|| anyhow::anyhow!("checksum failed"))?,
-        &src,
-        &dst,
-    );
-    echo.set_checksum(cksum);
-
-    let payload_len = echo.packet().len();
-    let mut ip = MutableIpv6Packet::owned(vec![0u8; 40 + payload_len])
-        .ok_or_else(|| anyhow::anyhow!("failed to create IPv6 packet"))?;
-    ip.set_version(6);
-    ip.set_payload_length(payload_len as u16);
-    ip.set_next_header(pnet_packet::ip::IpNextHeaderProtocols::Icmpv6);
-    ip.set_hop_limit(64);
-    ip.set_source(src);
-    ip.set_destination(dst);
-    ip.set_payload(echo.packet());
-
-    Ok(ip.consume_to_immutable().packet().to_vec())
-}
-
-fn is_echo_reply_v6(data: &[u8], expected_dst: Ipv6Addr) -> bool {
-    let Some(ip) = pnet_packet::ipv6::Ipv6Packet::new(data) else {
-        return false;
-    };
-    if ip.get_destination() != expected_dst {
-        return false;
-    }
-    if ip.get_next_header() != pnet_packet::ip::IpNextHeaderProtocols::Icmpv6 {
-        return false;
-    }
-    let Some(reply) = pnet_packet::icmpv6::echo_reply::EchoReplyPacket::new(ip.payload()) else {
-        return false;
-    };
-    if reply.get_icmpv6_type() != Icmpv6Types::EchoReply {
-        return false;
-    }
-    println!(
-        "  IPv6 reply: {} → {}, seq={}",
-        ip.get_source(),
-        ip.get_destination(),
-        reply.get_sequence_number(),
-    );
-    true
 }
